Add make lib-pkg
If a bootstrap compiler is built, make lib-pkg may be run before
configure to build all the required dependencies as findlib packages as
an alternative to running make lib-ext after configure.

The retrieval of the archives has much in common with the lib-ext
method, and the two mechanisms are designed to live in harmony in the
same Makefile. The patches directory is now ordered so that
src_ext/patches/foo are the patches for foo when built with lib-ext
(so as before) and src_ext/patches/foo.pkg are the patches for foo when
built with lib-pkg. Additionally, src_ext/patches/foo.common are used
with both mechanisms.

The URLs and MD5 checksums have been moved to src_ext/Makefile.sources
and the build instructions for each package are placed in
src_ext/Makefile.packages.

src_ext/Makefile has been updated to work without ../Makefile.config
and Makefile cookies are used to ensure that the clone and clone-pkg
targets correctly set-up the build directories.

make lib-pkg supports parallel make.

+++ .travis-ci.sh
@@ -26,7 +26,11 @@ install-bootstrap () {
 
 case "$TARGET" in
     prepare)
-        mkdir -p ~/local/bin
+      mkdir -p ~/local/bin
+
+      if [ "$COLD" = "1" ]; then
+        make compiler
+      else
         wget -q -O ~/local/bin/opam \
              "https://github.com/ocaml/opam/releases/download/$OPAMBSVERSION/opam-$OPAMBSVERSION-$(uname -m)-$(uname -s)"
         chmod a+x ~/local/bin/opam
@@ -35,21 +39,26 @@ case "$TARGET" in
             rvm --default use 2.3.3
             brew install "$EXTERNAL_SOLVER"
         fi
-        exit 0
-        ;;
+      fi
+      exit 0
+      ;;
     install)
+      if [ "$COLD" = "1" ]; then
+        make lib-pkg
+      else
         # Note: this part is cached, and must be idempotent
         # Re-init opam from scratch if the install fails
         if [ -d $OPAMBSROOT ]
         then install-bootstrap || { rm -rf $OPAMBSROOT; install-bootstrap; }
         else install-bootstrap
         fi
-        exit 0
-        ;;
+      fi
+      exit 0
+      ;;
     build)
-        ;;
+      ;;
     *)
-        echo "bad command $TARGET"; exit 1
+      echo "bad command $TARGET"; exit 1
 esac
 
 export OPAMYES=1
@@ -60,14 +69,16 @@ git config --global user.email "travis@example.com"
 git config --global user.name "Travis CI"
 
 ( # Run subshell in bootstrap root env to build
+  if [ "$COLD" != "1" ]; then
     export OPAMROOT="$OPAMBSROOT"
     eval $(opam config env)
 
     [ "$(ocaml -vnum)" = "$OCAML_VERSION" ] || exit 12
+  fi
 
     ./configure --prefix ~/local --with-mccs
 
-    if [ "$OPAM_TEST" != "1" ]; then make lib-ext; fi
+    if [ "$OPAM_TEST" != "1" -a "$COLD" != "1"  ]; then make lib-ext; fi
     make all
 
     rm -f ~/local/bin/opam
@@ -96,6 +107,9 @@ git config --global user.name "Travis CI"
         cd ~/build/opam-rt-*
         OPAMEXTERNALSOLVER="$EXTERNAL_SOLVER" make KINDS="local git" run
     else
+        if [ "$COLD" = "1" ]; then
+          export PATH=`pwd`/bootstrap/ocaml/bin:$PATH
+        fi
         # Test basic actions
         opam init
         eval $(opam config env)
